+++ apps/web/pages/[user]/embed.tsx
@@ -1,6 +1,7 @@
-import withEmbedSsr from "@lib/withEmbedSsr";
+"use client";
 
-import { getServerSideProps as _getServerSideProps } from "../[user]";
+import { getServerSideProps as _getServerSideProps } from "@lib/[user]/getServerSideProps";
+import withEmbedSsr from "@lib/withEmbedSsr";
 
 export { default } from "../[user]";
